Separate the robot controller from Primitives UNIT TEST PASSED * real.pkl recovered * put the RobotController under plb.engine directly * update the unit tests * update the YAML Config Parser
