+++ plb/config/default_config.py
@@ -26,6 +26,10 @@ _C.SIMULATOR.gravity = (0, -1, 0)
 # PRIMITIVES, i.e., Controller
 # ---------------------------------------------------------------------------- #
 _C.PRIMITIVES = list()
+# ---------------------------------------------------------------------------- #
+# ROBOTS, i.e., RobotController
+# ---------------------------------------------------------------------------- #
+_C.ROBOTS = list()
 
 # ---------------------------------------------------------------------------- #
 # Controller
