+++ tests/unit/test_robot_controller.py
@@ -2,23 +2,24 @@ from typing import Any, Dict
 
 import numpy as np
 import taichi
+import torch
 
-from plb.engine.primitive.primitives import RobotsControllers
-from plb.urdfpy import Robot, Link, FK_CFG_Type, Mesh
+from plb.engine.robots import RobotsController
+from plb.urdfpy import DiffRobot, Link, FK_CFG_Type, Mesh
 
 taichi.init()
 
 def test_deflatten_robot_actions():
-    robot = Robot.load('tests/data/ur5/ur5.urdf')
+    robot = DiffRobot.load('tests/data/ur5/ur5.urdf')
     robotActionDim = sum((
         joint.action_dim for joint in robot.actuated_joints
     ))
     envAction = [np.random.random() for _ in range(robotActionDim)]
-    jointActions = RobotsControllers._deflatten_robot_actions(robot, envAction)
+    jointActions = RobotsController._deflatten_robot_actions(robot, envAction)
     assert np.all(np.equal(jointActions, envAction))
 
     envAction = [0.33, 0.59, 0.01] + envAction
-    jointActions = RobotsControllers._deflatten_robot_actions(robot, envAction[3:])
+    jointActions = RobotsController._deflatten_robot_actions(robot, envAction[3:])
     assert np.all(np.equal(jointActions, envAction[3:]))
 
     class FakeJoint:
@@ -34,7 +35,7 @@ def test_deflatten_robot_actions():
             ]
     
     envAction = [0.33, 0.59, 0.01, 0.1, 0.22, 0.33, 0.44, 0.55, 0.66]
-    jointActions = RobotsControllers._deflatten_robot_actions(FakeRobot(), envAction[3:])
+    jointActions = RobotsController._deflatten_robot_actions(FakeRobot(), envAction[3:])
     assert len(jointActions) == 3 and \
         jointActions[0] == 0.1 and \
         len(jointActions[1]) == 2 and jointActions[1][0] == 0.22 and jointActions[1][1] == 0.33 and\
@@ -42,8 +43,8 @@ def test_deflatten_robot_actions():
         jointActions[2][2] == 0.66, f"{envAction} => {jointActions}"
 
 def test_single_robot():
-    rc = RobotsControllers()
-    robot = Robot.load('tests/data/ur5/ur5.urdf')
+    rc = RobotsController()
+    robot = DiffRobot.load('tests/data/ur5/ur5.urdf')
 
 
     # test append robot
@@ -57,14 +58,14 @@ def test_single_robot():
         f"got {linkCnt} Primitives, but there are {truePrimitiveCnt} in the RC"
     assert 1 == len(rc.robots),\
         f"got {len(rc.robots)} robots in the RC, but expecting 1"
-    assert 1 == len(rc.link_2_primtives),\
-        f"got {len(rc.link_2_primtives)} link_2_primitive in the RC, but expecting 1"
+    assert 1 == len(rc.link_2_primitives),\
+        f"got {len(rc.link_2_primitives)} link_2_primitive in the RC, but expecting 1"
     for linkName, link in robot.link_map.items():
         if any((
             not isinstance(collision.geometry.geometry, Mesh)
             for collision in link.collisions
         )):
-            assert linkName in rc.link_2_primtives[0],\
+            assert linkName in rc.link_2_primitives[0],\
                 f"{linkName} of the loaded robot not in rc.link_2_primitive"
 
     # test append action dims
@@ -83,8 +84,11 @@ def test_single_robot():
         f" but expecting [0, 3, 6, 9, {robotActionDim}]"
     
     # test set robot actions
-    envAction = [0.33, 0.66, 0.72] + [np.random.random() for _ in range(robotActionDim)]
-    rc.set_robot_actions(envAction, 3)
+    envAction = [0.33, 0.66, 0.72] + [
+        torch.rand((1,), device='cuda', dtype=torch.float64, requires_grad=True)
+        for _ in range(robotActionDim)
+    ]
+    rc.set_robot_actions(1, envAction, 3)
     poseA: Dict[Link, Any] = robot._current_cfg
     robot.link_fk(envAction[3:], cfgType=FK_CFG_Type.angle)
     poseB: Dict[Link, Any] = robot._current_cfg
@@ -103,9 +107,9 @@ def test_single_robot():
 
 
 def test_dual_robot():
-    rc = RobotsControllers()
-    robotA = Robot.load('tests/data/ur5/ur5.urdf')
-    robotB = Robot.load('tests/data/ur5/ur5.urdf')
+    rc = RobotsController()
+    robotA = DiffRobot.load('tests/data/ur5/ur5.urdf')
+    robotB = DiffRobot.load('tests/data/ur5/ur5.urdf')
     
     linkCnt = sum((
         1 for _ in rc.append_robot(robotA)
@@ -121,21 +125,21 @@ def test_dual_robot():
         f"got {linkCnt} Primitives, but there are {truePrimitiveCnt} in the RC"
     assert 2 == len(rc.robots),\
         f"got {len(rc.robots)} robots in the RC, but expecting 2"
-    assert 2 == len(rc.link_2_primtives),\
-        f"got {len(rc.link_2_primtives)} link_2_primitive in the RC, but expecting 2"
+    assert 2 == len(rc.link_2_primitives),\
+        f"got {len(rc.link_2_primitives)} link_2_primitive in the RC, but expecting 2"
     for linkName, link in robotA.link_map.items():
         if any((
             not isinstance(collision.geometry.geometry, Mesh)
             for collision in link.collisions
         )):
-            assert linkName in rc.link_2_primtives[0],\
+            assert linkName in rc.link_2_primitives[0],\
                 f"{linkName} of the loaded robot not in rc.link_2_primitive"
     for linkName, link in robotB.link_map.items():
         if any((
             not isinstance(collision.geometry.geometry, Mesh)
             for collision in link.collisions
         )):
-            assert linkName in rc.link_2_primtives[1],\
+            assert linkName in rc.link_2_primitives[1],\
                 f"{linkName} of the loaded robot not in rc.link_2_primitive"
 
     action_dims = [0]
@@ -154,8 +158,11 @@ def test_dual_robot():
         f"after appending robot's action dims, the action_dims become {action_dims},"+\
         f" but expecting [0, 3, 6, 9, {robotActionDim}, {robotActionDim}]"
     
-    envAction = [0.33, 0.66, 0.72] + [np.random.random() for _ in range(robotActionDim * 2)]
-    rc.set_robot_actions(envAction, 3)
+    envAction = [0.33, 0.66, 0.72] + [
+        torch.rand((1,), device='cuda', dtype=torch.float64, requires_grad=True)
+        for _ in range(robotActionDim * 2)
+    ]
+    rc.set_robot_actions(1, envAction, 3)
     poseA: Dict[Link, Any] = robotB._current_cfg
     robotB.link_fk(envAction[3 + robotActionDim:], cfgType=FK_CFG_Type.angle)
     poseB: Dict[Link, Any] = robotB._current_cfg
